Backend/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
PyJWT>=2.3.0

# LangChain and LLM integration (pinned for compatibility)
langchain>=0.3.0
langchain-core>=0.3.41
langchain-community>=0.3.19
youtube_search>=2.1.2

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pymongo>=4.3.0
# Email validation required by pydantic
email-validator>=1.3.0
youtube-transcript-api>=0.6.0
duckduckgo-search>=2.7.8

# Retry and error handling
tenacity>=8.2.3

# Code execution and utilities
docker>=6.0.0
streamlit>=1.28.0
streamlit-ace>=0.1.1

# API clients
httpx>=0.25.0
requests>=2.31.0

# Utilities
python-multipart>=0.0.5
uuid>=1.30
typing-extensions>=4.8.0
arxiv>=2.1.3
pymupdf>=1.25.3